requirements.txt
Django==5.2.5
Pillow==11.1.0
django-crispy-forms==2.4
crispy-bootstrap5==2025.6
python-decouple==3.8
dj-database-url==3.0.1
psycopg==3.2.9
